reformed project in a modular way. url building is yet to be fixed
