+++ ytd_web/__init__.py
@@ -1,0 +1,10 @@
+# from flask import Blueprint
+
+# home_page = Blueprint('home_page', 'ytd_web.home')
+# search_page = Blueprint('search_page', "ytd_web.search")
+# browse_page = Blueprint('browse_page', "ytd_web.browse")
+# select_page = Blueprint('select_page', "ytd_web.select")
+# multimedia_page = Blueprint('multimedia_page', "ytd_web.multimedia")
+# get_miltimedia_page = Blueprint('get_multimedia_page', "ytd_web.get_multimedia")
+# download_page = Blueprint('download_page', "ytd_web.download")
+# error_page = Blueprint('error_page', "ytd_web.error")
